Add files via upload
This my Laptops SSD. 
This forced me to transitioning to Google Colab in very little time. 
I do have the scripts, but I am struggling currently with setting up this repo, especially given all the large files. 
I do have the classes mentioned in the Skripsie. 

This repo will not be abandoned, and streamline the process for the next student.